Move some stuff in the right packages
- This will help split stuff in smaller, better tested packages
- This moves some stuff like the traefik command to package `cmd`
