+++ server/adapters.go
@@ -1,7 +1,4 @@
-/*
-Copyright
-*/
-package main
+package server
 
 import (
 	"net/http"
